comapny_management/requirement.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.3
djangorestframework==3.14.0
et-xmlfile==1.1.0
numpy==1.24.4
openpyxl==3.1.2
pandas==2.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
